Move funcs of resource_usage_gatherer.go to e2e/scheduling/gatherer
